Cleaning up unnecessary stuff and adding some blocks to further work on

+++ hungergames_common/com/eiscube/hungergames/items/ItemWeaponBow.java
@@ -2,7 +2,6 @@ package com.eiscube.hungergames.items;
 
 import com.eiscube.hungergames.HungerGames;
 
-import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.entity.player.EntityPlayer;
@@ -41,10 +40,10 @@ public class ItemWeaponBow extends ItemBow {
 
         if (flag || par3EntityPlayer.inventory.hasItem(Item.arrow.itemID))
         {
-            float f = (float)j / 20.0F;
+            float f = j / 20.0F;
             f = (f * f + f * 2.0F) / 3.0F;
 
-            if ((double)f < 0.1D)
+            if (f < 0.1D)
             {
                 return;
             }
@@ -66,7 +65,7 @@ public class ItemWeaponBow extends ItemBow {
 
             if (k > 0)
             {
-                entityarrow.setDamage(entityarrow.getDamage() + (double)k * 0.5D + 0.5D);
+                entityarrow.setDamage(entityarrow.getDamage() + k * 0.5D + 0.5D);
             }
 
             int l = EnchantmentHelper.getEnchantmentLevel(Enchantment.punch.effectId, par1ItemStack);
